More HTTP fault tolerance
Storing HTTP headers as lowercase internally, to allow for partially
malformed headers to be still understood. Removing check for HTTP
version, and also “intelligently deducting” the initial HTTP-Request
line, if it is only partially, or incorrectly sent from client. The
idea is to create a highly fault tolerant server, allowing for at least
partially malformed requests to still be understood. In addition to not
let a partially malformed request compromise the server in any ways.