Significantly improved "HTTP cloaking" support.
Support for using any characters [1 10] ‘\n’ and any range from [11 20]
as ‘\r’, in string_helper::get_line, which dramatically improves its
HTTP cloaking abilities, since any adversaries cannot even look for a
CR/LF sequence to understand if it has found a match on a brute force
decryption attack of a message.

In addition the string_helper::get_line now completely discards
anything in the range [21 32> and anything in the range [127 256>,
which allows for client to put “garbage data” into the request in the
envelope. (headers and HTTP-Request line)

More to be done, especially in the match_condition class, but the rough
idea is getting there.

Stuffed the boost::asio::streambuf into the connection class, to
support pipelining. Fixed some bugs in the implementation of timeouts
and usage of deadline_timer. Still some funny exceptions, which I think
is due to not “posting connection close as work” to io_service,
somehow. More investigation to be done …