TRACE method handler.
Implemented support for TRACE method, such that you can create a
request, and see the request you created, returned as the body from
server, the way the server would interpret and understand it.

Cleaned up implementation for HTTP headers and GET parameters in the
request_envelope class, to make it more intuitive in usage. Changed
implementation from std::map to std::vector<tuple<string, string> >, to
make it easier to iterate over name/value.

Implemented support for HTTP headers spanning multiple lines, with a SP
and/or TAB starting the next line.

Fixed bug that would make “?” become a part of the URI when parsing
HTTP envelope.