SSL mostly in place
Need some refactoring and checking up that things are as they should
be, but largely, the ghost of it, is in place.

Creates an abstract bas class called “rosetta_socket”, which implements
in its two different derived classes, all socket operations, to be able
to use both SSL and normal plain sockets in an abstract way. Connection
contains an instance of this class now.

Two configuration properties are there now for .crt file and .key file.