Fixed a segmentation fault
Which I think was because of me misunderstanding of how the boost Casio
strand class worked.

Started implementing authentication, and moving on to authorization.

Fixed a bug when redirecting to secure version request, which would
remove “index.html” when redirecting, which shouldn’t be done anymore,
due to being able to GET folder contents.

Ne exception type; security_exception, which is means for violations of
what’s considered an attempt at hacking into the server.