Finally seems to have eliminated dangling pointers
It seemed that connection and/or socket, would be destroyed, before all
pieces of code, referencing them would finish. This is no longer the
case, since we keep connection, and socket, alive until everything is
cleanly finished. Even when a deadline_timer stops the connection, and
or the client drops the connection.